Move random gen logic into token file
