+++ token.go
@@ -1,10 +1,14 @@
 package password
 
 import (
+	"crypto/rand"
+	"encoding/base64"
 	"time"
 
 	jwt "github.com/dgrijalva/jwt-go"
 )
+
+var signingKey = genRandBytes()
 
 // GenToken generates a new JSON web token.
 func GenToken(id string) (string, error) {
@@ -21,4 +25,28 @@ func GenToken(id string) (string, error) {
 	}
 
 	return tokStr, nil
+}
+
+// SetSigningKey allows you to override the default HMAC signing key with one
+// of your own. Every time this package is imported, a signing key is set
+// randomly. That means that in between restarts, a new key is set, so you'd
+// no longer be able to verify JSON web tokens created with that key. In order
+// to reuse the signing key, you must set it yourself. Just call this function
+// before creating any tokens, and you'll be good to go.
+func SetSigningKey(key []byte) {
+	signingKey = key
+}
+
+func genRandBytes() []byte {
+	// Use 32 bytes (256 bits) to satisfy the requirement for the HMAC key
+	// length.
+	b := make([]byte, 32)
+	_, err := rand.Read(b)
+	if err != nil {
+		// If this errors, it means that something is wrong the system's
+		// CSPRNG, which indicates a critical operating system failure. Panic
+		// and crash here
+		panic(err)
+	}
+	return []byte(base64.URLEncoding.EncodeToString(b))
 }
